pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django-cors-headers>=4.9.0",
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "supabase>=2.19.0",
    "werkzeug>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "email-validator>=2.3.0",
    "gunicorn>=23.0.0",
]
